pyproject.toml
[project]
name = "building_lod2_builder_heat"
version = "0.1.0"
description = ""
authors = [
    {name = "fukuchidaisuke",email = "fukuchidaisuke@realglobe.jp"}
]
readme = "README.md"
license = {text = "GPL-3.0-or-later"}
requires-python = ">=3.13,<4"
dependencies = [
    "laspy (>=2.6.1,<3.0.0)",
    "lazrs (>=0.7.0,<0.8.0)",
    "numpy (>=2,<2.3.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "pyproj (>=3.7.1,<4.0.0)",
    "rasterio (>=1.4.3,<2.0.0)",
    "typer[all] (>=0.16.0,<0.17.0)",
    "heat @ git+https://github.com/realglobe-Inc/heat.git@2025-dev",
    # # 以下はHEAT開発時に一時的に
    # "cython (>=3.1.2,<4.0.0)",
    # "einops (>=0.8.1,<0.9.0)",
    # "matplotlib (>=3.10.3,<4.0.0)",
    # "numpy (>=2,<2.3.0)",
    # "opencv-python (>=4.12.0.88,<5.0.0.0)",
    # "pillow (>=11.3.0,<12.0.0)",
    # "pyproj (>=3.7.1,<4.0.0)",
    # "scikit-learn (>=1.7.0,<2.0.0)",
    # "scipy (>=1.16.0,<2.0.0)",
    # "shapely (>=2.1.1,<3.0.0)",
    # "torch (>=2.7.1,<3.0.0)",
    # "torchvision (>=0.22.1,<0.23.0)",
    # "heat @ file:///path/to/local/heat",
]


[project.urls]
"torch-stable" = "https://download.pytorch.org/whl/torch_stable.html"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
extract-roofline = "building_lod2_builder_heat.commands.extract_roofline.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
ninja = "^1.11.1.4"
pytest = "^8.4.1"
scipy-stubs = "^1.16.0.2"

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests",
]
